stock_service/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.0
drf-yasg==1.21.7
requests==2.31.0
Pillow==10.1.0
uuid6
django-extensions
